Adding climate/gridding module
The gridding module takes a N*M array of data where N
is the number of data points and M divided into the axes
A and the data D such that A+D=M.

Each axis D is associated with an output axes that determines
the upper bin edges (or number of bins with min and max values)
and optional missing_data values (or masked arrays). The data
is binned into each of the D dimensions and the average, variance, and
count of each bin is calculated for each entry in A.

Entries in A can contain data from D, but must be explicitly included
(mean axes values are not calculated by default).

Optionally, the calculation can be done using a sparse matrix algorithm
to reduce the waste if the number of output bins is large, the input
data is clustered, or a sorted 1 dimensional version of the multi-dimensional
grid is required (sorted by the order of the axes).

This code was developed at Ashima Research and has lived in the pyJura package for
a while.